+++ brownie/parallel.py
@@ -20,7 +20,7 @@ except ImportError:
         if sys.platform == 'win32':
             try:
                 return int(os.environ['NUMBER_OF_PROCESSORS'])
-            except ValueError, KeyError:
+            except (ValueError, KeyError):
                 # value could be anything or not existing
                 pass
         if sys.platform in ('bsd', 'darwin'):
@@ -33,7 +33,7 @@ except ImportError:
             cpu_count = os.sysconf('SC_NPROCESSORS_ONLN')
             if cpu_count >= 1:
                 return cpu_count
-        except AttributeError, ValueError:
+        except (AttributeError, ValueError):
             # availability is restricted to unix
             pass
         if default is not None: