+++ CHANGELOG.rst
@@ -3,6 +3,14 @@ Changelog
 This log contains a list of bug fixes and added, modified or even removed
 features.
 
+0.1.1
+-----
+*Currently in development*
+
+- Fixed a :exc:`KeyError` and a :exc:`ValueError` which could occur
+  by calling :func:`brownie.parallel.get_cpu_count` on Windows or Linux
+  respectively.
+
 0.1 - Fairy Land
 ----------------